Refactor components into individual files & code style changes
